+++ game-baldursgate3/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.7] - UNRELEASED
+
+- Added activity notification when importing load order
+- Fixed importing missing mods name comparison
+- Removed some spamming console logs   
+
 ## [1.3.6] - UNRELEASED
 
 - Load Order now uses pak filename to uniquely identify